Crate audio

Structs

AudioBuffer
A container for audio samples and important attributes.

Enums

AudioError
An enumeration for reporting audio errors

AudioFormat
All supported audio formats.

Codec
All supported audio codecs.

SampleOrder
The channel ordering of audio Samples.

Traits

FromSample
Converts a Sample to another type.

ToSample
Converts a type to a Sample.

Functions

load
Loads the audio from a reader into memory.

open
Opens and loads the audio file into memory from a Path.

save
Saves an AudioBuffer to a Path.

save_as
Saves an AudioBuffer to a Path using a specified Codec.

write
Buffers and writes an AudioBuffer to a writer using a specified AudioFormat.

write_as
Buffers and writes an AudioBuffer to a writer using a specified AudioFormat and Codec.

Type Definitions

AudioResult
Result type of an audio encoding or decoding process

Sample
Audio sample.