Function audio::save

pub fn save(path: &Path, audio: &AudioBuffer) -> AudioResult<()>

Saves an AudioBuffer to a Path.

The necessary encoder is determined by the Path file extension and uses the default codec of the AudioFormat. An AudioError is returned if the file type is not supported or if an error occurred in the encoding process.