Struct audio::AudioBuffer

pub struct AudioBuffer {
    pub sample_rate: u32,
    pub channels: u32,
    pub samples: Vec<Sample>,
}

A container for audio samples and important attributes.

Fields

sample_rate

Number of samples per second

channels

Number of channels

samples

Decoded audio samples

Methods

impl AudioBuffer

fn new(sample_rate: u32, channels: u32) -> Self

Creates a new, empty AudioBuffer.

fn from_samples(sample_rate: u32, channels: u32, samples: Vec<Sample>) -> Self

Creates a new AudioBuffer using the given Samples.

fn from_bytes(sample_rate: u32, channels: u32, bytes: &[u8], codec: Codec) -> AudioResult<Self>

Creates an AudioBuffer from a set of bytes using a Codec.

Bytes are interpreted using a Codec and are passed to the new AudioBuffer set with the specified sample rate and number of channels.

fn duration(&self) -> usize

The duration of the audio in milliseconds.